pyproject.toml
[project]
name = "inventiobackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-filter>=25.1",
    "django-stubs>=5.1.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djangorestframework-stubs>=3.15.3",
    "drf-spectacular>=0.28.0",
    "markdown>=3.8",
    "pip>=25.0.1",
    "psycopg>=3.2.6",
    "python-decouple>=3.8",
    "tzdata>=2025.2",
]
